requirements.txt
asgiref==3.4.1
Django==4.0.5
djangorestframework==3.13.1
cryptography==37.0.2
django-rest-knox==4.2.0
psycopg2==2.9.3
gunicorn==20.1.0